Minor categorization metrics refactoring
