+++ freediscovery/metrics.py
@@ -8,6 +8,8 @@ from __future__ import unicode_literals
 import numpy as np
 from .cluster.utils import _dbscan_unique2noisy
 
+from sklearn.metrics.base import UndefinedMetricWarning
+
 # Information retrieval specific metrics
 # See scikit learn metrics for more general ones
 
@@ -183,4 +185,69 @@ def _scale_cosine_similarity(x, metric='cosine'):
     return x
 
 
+
+def categorization_score(idx_ref, Y_ref, idx, Y):
+    """ Calculate the efficiency scores """
+    # This function should be deprecated
+    # An equivalent functionally should be achieved with a
+    # more general freediscovery.metrics module
+    import warnings
+    from sklearn.metrics import (precision_score, recall_score, f1_score,
+            roc_auc_score, average_precision_score)
+    threshold = 0.0
+
+    idx = np.asarray(idx, dtype='int')
+    idx_ref = np.asarray(idx_ref, dtype='int')
+    Y = np.asarray(Y)
+    Y_ref = np.asarray(Y_ref)
+
+    idx_out = np.intersect1d(idx_ref, idx)
+    if not len(idx_out):
+        return {"recall_score": -1, "precision_score": -1, 'f1': -1, 'auc_roc': -1,
+                'average_precision': -1}
+
+    # sort values by index 
+    order_ref = idx_ref.argsort()
+    idx_ref = idx_ref[order_ref]
+    Y_ref = Y_ref[order_ref]
+
+    order = idx.argsort()
+    idx = idx[order]
+    Y = Y[order]
+
+    # find indices that are in both the reference and the test dataset
+    mask_ref = np.in1d(idx_ref, idx_out)
+    mask = np.in1d(idx, idx_out)
+
+    Y_ref = Y_ref[mask_ref]
+    Y = Y[mask]
+    Y_bin = (Y > threshold)
+
+    n_classes = len(np.unique(Y_ref))
+
+    if n_classes != 2:
+        opts = {"average": 'micro'}
+    else:
+        opts = {}
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", category=UndefinedMetricWarning)
+
+        m_recall_score = recall_score(Y_ref, Y_bin, **opts)
+        m_precision_score = precision_score(Y_ref, Y_bin, **opts)
+        m_f1_score = f1_score(Y_ref, Y_bin, **opts)
+    if n_classes == 2:
+        m_roc_auc = roc_auc_score(Y_ref, Y)
+        m_average_precision = average_precision_score(Y_ref, Y)
+        m_recall_at_20p = recall_at_k_score(Y_ref, Y, k=0.2)
+    else:
+        # not defined for non binary categorization
+        m_roc_auc = np.nan
+        m_average_precision = np.nan
+        m_recall_at_20p = np.nan
+
+    return {"recall": m_recall_score, "precision": m_precision_score,
+            "f1": m_f1_score, 'roc_auc': m_roc_auc,
+            'average_precision': m_average_precision,
+            'recall_at_20p': m_recall_at_20p}
