+++ freediscovery/utils.py
@@ -12,13 +12,14 @@ from contextlib import contextmanager
 import pandas as pd
 import numpy as np
 import uuid
-try:  # sklearn v0.17
-    from sklearn.exceptions import UndefinedMetricWarning
-except ImportError:  # v0.18
-    from sklearn.metrics.base import UndefinedMetricWarning
+from sklearn.metrics.base import UndefinedMetricWarning
 
 from .exceptions import (DatasetNotFound, ModelNotFound, InitException,
                             WrongParameter)
+
+
+# this is wrong, and should be eventually replaced
+INT_NAN = -99999
 
 @contextmanager
 def _silent(stream='stderr'):
@@ -28,70 +29,6 @@ def _silent(stream='stderr'):
     yield
     setattr(sys, stream, stderr)
 
-
-INT_NAN = -99999
-
-
-def categorization_score(idx_ref, Y_ref, idx, Y):
-    """ Calculate the efficiency scores """
-    # This function should be deprecated
-    # An equivalent functionally should be achieved with a
-    # more general freediscovery.metrics module
-    import warnings
-    from sklearn.metrics import (precision_score, recall_score, f1_score,
-            roc_auc_score, average_precision_score)
-    threshold = 0.0
-
-    idx = np.asarray(idx, dtype='int')
-    idx_ref = np.asarray(idx_ref, dtype='int')
-    Y = np.asarray(Y)
-    Y_ref = np.asarray(Y_ref)
-
-    idx_out = np.intersect1d(idx_ref, idx)
-    if not len(idx_out):
-        return {"recall_score": -1, "precision_score": -1, 'f1': -1, 'auc_roc': -1,
-                'average_precision': -1}
-
-    # sort values by index 
-    order_ref = idx_ref.argsort()
-    idx_ref = idx_ref[order_ref]
-    Y_ref = Y_ref[order_ref]
-
-    order = idx.argsort()
-    idx = idx[order]
-    Y = Y[order]
-
-    # find indices that are in both the reference and the test dataset
-    mask_ref = np.in1d(idx_ref, idx_out)
-    mask = np.in1d(idx, idx_out)
-
-    Y_ref = Y_ref[mask_ref]
-    Y = Y[mask]
-    Y_bin = (Y > threshold)
-
-    n_classes = len(np.unique(Y_ref))
-
-    if n_classes != 2:
-        opts = {"average": 'micro'}
-    else:
-        opts = {}
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore", category=UndefinedMetricWarning)
-
-        m_recall_score = recall_score(Y_ref, Y_bin, **opts)
-        m_precision_score = precision_score(Y_ref, Y_bin, **opts)
-        m_f1_score = f1_score(Y_ref, Y_bin, **opts)
-    if n_classes == 2:
-        m_roc_auc = roc_auc_score(Y_ref, Y)
-        m_average_precision = average_precision_score(Y_ref, Y)
-    else:
-        m_roc_auc = np.nan # ROC not defined in this case
-        m_average_precision = np.nan
-
-    return {"recall": m_recall_score, "precision": m_precision_score,
-            "f1": m_f1_score, 'roc_auc': m_roc_auc,
-            'average_precision': m_average_precision }
 
 def _rename_main_thread():
     """
